Support for titleAttr as an optional reference
I have added support for making a tittleAttr an optional reference or a
simple attribute.  This is primarily done by adding a prepareEvents
method "in front of" the createEvents method that does a
mx.processor.get() to pull in any referred attributes.

Note: I have also added two minor features that are supported by the
underlying fullcalendar.js, but not currently  supported in the Mendix
Calendar widget.

The first is to support the axisFormat property that controls the time
labels along the left side of day and agenda views.  In particular, this
needs to be adjusted to HH:mm if one prefers to use a 24-hour clock
instead of the default hh(:mm)tt format string.

Second, I have exposed the slotMinutes property that allows the time
increment of day and agenda calenders to be altered from the default of
30 minutes.  For example, in a law office, one might want to support
time increments of either 10 or 15 minutes for billing purposes.
Reserving tennis courts using the Calendar widget might wish to have a
minimum calendar increment of 60 minutes.